webempresa/requirements.txt
Django~=4.2.6
django-ckeditor~=6.7.0